feat(landing): apex pricing + crawlable nav (mirror 9.2.3/9.2.4) — v10.0.3
Forward-ports the maintenance-line landing fixes to 10.x. Supersedes
PricingSeoListener with LandingSeoListener:

- Every page's server HTML gets a <noscript> nav to all public pages, so a
  link-following crawler can discover /pricing (the nav is otherwise
  Vaadin-rendered and absent from server HTML).
- The home page shows a compact "Simple, per-member pricing" band (existing
  pricing card system; $7/mo, $60/yr per member, 60-day free trial, taxes at
  checkout), and the pricing JSON-LD + <noscript> summary is injected on "/" as
  well as /pricing — so the apex (the URL an MoR registers and checks) carries
  the price in raw HTML.

Bump VERSION + 5 poms to 10.0.3, add CHANGELOG entry. 67 landing tests pass.
