+++ tnra-app/pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>com.afitnerd</groupId>
     <artifactId>tnra-app</artifactId>
-    <version>10.0.2</version>
+    <version>10.0.3</version>
     <packaging>jar</packaging>
 
     <name>tnra-app</name>
